+++ elmyra.ip.access.epo/CHANGES.rst
@@ -21,6 +21,7 @@ development
 - Display both (docdb) and (epodoc) numbers at application reference
 - depa.tech: Highlighting for expert search
 - Use recent browser-like User-Agent across the board
+- Improve patent number normalization and usage
 
 2017-04-06 0.134.1
 ==================